[ENH]  Elbow class selection/pairwise dimension selection for multivariate time series classification  (#2941)
* Add ECP and Kmeas based method

* Update ECS, ECP and CCS method

* Update ECP, ECS, and CCS

* Update ECS, ECP and KMeans channel selection method

* Update example notebook
